+++ react/ui-components/src/constants/images/images.js
@@ -1,0 +1,7 @@
+const IMAGES = {
+  DIGIT_DARK: "https://egov-dev-assets.s3.ap-south-1.amazonaws.com/digit.png",
+  DIGIT_LIGHT:
+    "https://egov-uat-assets.s3.ap-south-1.amazonaws.com/hcm/mseva-white-logo.png",
+};
+
+export { IMAGES };