+++ react/ui-components/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2-beta.11] - 2024-07-04
+### New Changes
+- Added ActionBar, Header, ActionButton variant for Button
+- Updated Timeline Molecule 
+
 ## [0.0.2-beta.10] - 2024-06-28
 ### New Changes
 - Added Card Variants